Fix NPE in tests
BackupRef.get() is failing because the static cfg member is null.

As a workaround, make getTimestampBranch public and use that
directly. Since there are no tests for non-timestamp implementation
yet anyway, this should be OK.

Tests for the non-timestamp implementation, and correction in the
implementation of get() can be added in a follow-up commit.

Change-Id: I644db55f5c1cb043479ca3ea9b10b3b3c642a069